Fixing build warnings (#707)
* Fixing build warnings generated from deprecated function call.

+++ swri_roscpp/test/topic_service_test.cpp
@@ -102,7 +102,7 @@ namespace swri_roscpp
     }
 
     rclcpp::Node::SharedPtr node_;
-    int call_count_;
+    size_t call_count_;
     bool error_;
     bool is_running_;
   };
